+++ Ufo_pierwsza_wersja/constants.h
@@ -2,8 +2,8 @@
 #include <SFML/Graphics.hpp>
 
 namespace constants {
-	constexpr int window_width = 1920;  
-	constexpr int window_height = 1080;
+	constexpr float window_width = 1920;  
+	constexpr float window_height = 1080;
 	constexpr float missile_velocity_x = 0;
 	constexpr float missile_velocity_y = 10.0f;
 	constexpr int MISSILE_GENERATION_PERIOD = 3000;
